% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnett.R
\docType{package}
\name{garnett-package}
\alias{garnett}
\alias{garnett-package}
\title{garnett}
\description{
Garnett faciliates automated cell type classification from 
    single-cell expression (and other genomic) data. Garnett works by taking 
    single-cell data, along with a cell type definition (marker) file, and 
    training a regression-based classifier. Once a classifier is trained for a 
    tissue/sample type, it can be applied to classify future datasets from 
    similar tissues.
}
