% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnett_classifier.R
\docType{class}
\name{garnett_classifier}
\alias{garnett_classifier}
\alias{garnett_classifier-class}
\title{The garnett_classifier class}
\description{
Classifies cells according to a hierarchy of types.
}
\details{
Classifies cells according to a hierarchy of types via user-defined gating
functions.
}
\section{Slots}{

 \describe{
   \item{\code{classification_tree}:}{Object of class \code{"igraph"}}
   \item{\code{cell_totals}:}{Object of class \code{"numeric"}}
   \item{\code{gene_id_type}:}{Object of class \code{"character"}}
   \item{\code{references}:}{Object of class \code{"list"}}
 }
}

