% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_rule.R
\docType{class}
\name{gene_rule-class}
\alias{gene_rule-class}
\title{gene_rule class}
\description{
Class definition for gene_rule, which is a container to hold a gene name
along with upper and lower expression bounds.
}
\section{Slots}{

\describe{
\item{\code{gene_name}}{character. The name of the gene}

\item{\code{lower}}{numeric. Lower bound of expression - same units as CDS object.}

\item{\code{upper}}{numeric. Upper bound of expression - same units as CDS object.}
}}

