% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_classifier_references}
\alias{get_classifier_references}
\title{Retrieve marker references from garnett_classifier}
\usage{
get_classifier_references(classifier, cell_type = NULL)
}
\arguments{
\item{classifier}{garnett_classifier created using train_cell_classifier.}

\item{cell_type}{Cell type name or \code{NULL}. References for which cell
type should be printed? If \code{NULL}, all are printed.}
}
\value{
List of references included when garnett_classifier was trained.
}
\description{
Retrieve marker references from garnett_classifier
}
\examples{
data(test_classifier)
get_classifier_references(test_classifier)

}
