% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_feature_genes}
\alias{get_feature_genes}
\title{Extract feature genes}
\usage{
get_feature_genes(classifier, node = "root", convert_ids = FALSE, db = NULL)
}
\arguments{
\item{classifier}{Trained garnett_classifier - output from
\code{\link{train_cell_classifier}}.}

\item{node}{Character. The name of the parent node of the multinomial
classifier you would like to view features for. If top level, use "root".}

\item{convert_ids}{Logical. Should classifier IDs be converted to SYMBOL?}

\item{db}{Bioconductor AnnotationDb-class package for converting gene IDs.
For example, for humans use org.Hs.eg.db. See available packages at
\href{http://bioconductor.org/packages/3.8/data/annotation/}{Bioconductor}.
If \code{convert_ids = FALSE}, db can be \code{NULL}.}
}
\value{
A data.frame of coefficient values for each gene with non-zero
 coefficients in the classifier.
}
\description{
Extract the genes chosen as features in cell type classification from a
trained garnett_classifier
}
\examples{
library(org.Hs.eg.db)
data(test_classifier)
featuresdf <- get_feature_genes(test_classifier, db=org.Hs.eg.db)
featuresdf2 <- get_feature_genes(test_classifier,
                                 convert_ids = FALSE,
                                 node = "T cells")

}
