% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_markers}
\alias{plot_markers}
\title{Plot marker metrics}
\usage{
plot_markers(marker_check_df, amb_marker_cutoff = 0.5, label_size = 2)
}
\arguments{
\item{marker_check_df}{Marker check data.frame - output of check_markers.}

\item{amb_marker_cutoff}{Numeric. Cutoff at which to label ambiguous markers.
Default 0.5.}

\item{label_size}{Numeric, size of the text labels for ambiguous markers and
unplotted markers.}
}
\value{
A ggplot object of the marker plot.
}
\description{
This function takes as input the output of the \code{\link{check_markers}}
function and generates a plot to visualize the most important metrics.
}
\examples{
library(org.Hs.eg.db)

marker_file_path <- system.file("extdata", "pbmc_test.txt",
                                package = "garnett")
data(test_cds)
marker_check <- check_markers(test_cds,
                              marker_file_path,
                              db=org.Hs.eg.db,
                              cds_gene_id_type = "SYMBOL",
                              marker_file_gene_id_type = "SYMBOL")

plot_markers(marker_check)



}
