% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_cds}
\alias{test_cds}
\title{Small test CDS object}
\format{A CDS object with 32738 features and 800 samples}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/}
}
\usage{
test_cds
}
\description{
A CDS object used for testing and demonstration. Derives from the 10x
Genomics V1 PBMC single-cell RNA-seq dataset.
}
\keyword{datasets}
