% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_classifier}
\alias{test_classifier}
\title{Small test garnett_classifier object}
\format{An object of class \code{garnett_classifier} of length 1.}
\usage{
test_classifier
}
\description{
A small test garnett_classifier trained using the \code{test_cds} object.
}
\keyword{datasets}
